Added button to container, email verification
QOL improvements

+++ loginAuth.js
@@ -20,28 +20,48 @@ const backToRoomService = document.getElementById('backToRoomService');
 const displayForm = document.getElementById('displayForm');
 const displayCancel = document.getElementById('displayCancel');
 const displayFinalize = document.getElementById('displayFinalize');
+const emailAuthRoom = document.getElementById('emailAuthRoom');
+const cancelAuth = document.getElementById('cancelAuth');
+const mainBackToRoomService = document.getElementById('mainBackToRoomService');
 const db = firebase.database().ref();
-const databaseRoom = firebase.database().ref(joinName);
 
 var name;
 var joinName;
 var joinPass;
 var userDisplayName;
+//var user = firebase.auth().currentUser;
 
 // Login
 
 buttonLogin.addEventListener('click', function(e) {
   const emailVal = email.value;
   const passVal = pass.value;
-  const auth = firebase.auth();
-  const promise = auth.signInWithEmailAndPassword(emailVal, passVal);
-  roomService.removeAttribute('hidden');
-  buttonLogout.removeAttribute('hidden');
-  loginForm.setAttribute('hidden', 'loginForm');
-  invalidText.setAttribute('hidden', 'invalidText');
+  const promise = firebase.auth().signInWithEmailAndPassword(emailVal, passVal);
 
-  promise.catch(function(e) {
-    console.log(e.message);
+  firebase.auth().onAuthStateChanged(function(user) {
+      firebase.auth().currentUser.reload();
+      if (firebase.auth().currentUser.emailVerified) {
+        if (firebase.auth().currentUser.displayName == null) { // No name
+          displayNameForm.removeAttribute('hidden');
+          roomService.setAttribute('hidden', 'roomService');
+          buttonLogout.setAttribute('hidden', 'buttonLogout');
+          loginForm.setAttribute('hidden', 'loginForm');
+          invalidText.setAttribute('hidden', 'invalidText');
+        } else { // Green light
+
+          displayNameForm.setAttribute('hidden', 'displayNameForm');
+          roomService.removeAttribute('hidden');
+          loginForm.setAttribute('hidden', 'loginForm');
+          console.log(name);
+          console.log(firebase.auth().currentUser.displayName);
+        }
+      } else { // click here to send another email.
+        emailAuthRoom.removeAttribute('hidden');
+        loginForm.setAttribute('hidden', 'loginForm');
+      }
+    promise.catch(function(e) {
+      console.log(e.message);
+    });
   });
 });
 
@@ -51,28 +71,46 @@ buttonSignUp.addEventListener('click', function(e) {
   const emailVal = email.value;
   const passVal = pass.value;
   const auth = firebase.auth();
-  if ((emailVal.includes('@') && emailVal.includes('.com'))) {
+  if ((emailVal.includes('@') && emailVal.includes('.'))) {
     firebase.auth().createUserWithEmailAndPassword(emailVal, passVal)
       .catch(function(error) {
         console.log(error);
       });
-
   } else {
     invalidText.removeAttribute('hidden');
   }
-  displayNameForm.removeAttribute('hidden');
-  loginForm.setAttribute('hidden', 'loginForm');
+  firebase.auth().onAuthStateChanged(function(user) {
+    if (!(firebase.auth().currentUser.emailVerified)) {
+      firebase.auth().currentUser.sendEmailVerification().then(function() {
+        loginForm.setAttribute('hidden', 'loginForm');
+        emailAuthRoom.removeAttribute('hidden');
+        console.log("email sent");
+      }, function(error) {
+        console.log(error);
+      });
+    }
+  });
 });
 
+cancelAuth.addEventListener('click', function(e) {
+  emailAuthRoom.setAttribute('hidden', 'emailAuthRoom');
+  loginForm.removeAttribute('hidden');
+});
+
+// Setting username
+
 displayFinalize.addEventListener('click', function(e) {
-  firebase.auth().currentUser.updateProfile({displayName: document.getElementById('displayForm').value});
+  var user = firebase.auth().currentUser;
+  user.updateProfile({
+    displayName: displayForm.value
+  });
+  console.log(user);
   roomService.removeAttribute('hidden');
   buttonLogout.removeAttribute('hidden');
   loginForm.setAttribute('hidden', 'loginForm');
   invalidText.setAttribute('hidden', 'invalidText');
   displayNameForm.setAttribute('hidden', 'displayNameForm');
-
-})
+});
 
 displayCancel.addEventListener('click', function(e) {
   var user = firebase.auth().currentUser;
@@ -84,7 +122,7 @@ displayCancel.addEventListener('click', function(e) {
   if (user) {
     user.delete();
   }
-})
+});
 
 // Log out
 
@@ -102,7 +140,7 @@ logOutInner.addEventListener('click', function(e) {
   buttonLogout.setAttribute('hidden', 'buttonLogout');
   container.setAttribute('hidden', 'container');
   loginForm.removeAttribute('hidden');
-})
+});
 
 // Back to menu
 
@@ -110,7 +148,14 @@ if (backToRoomService) {
   backToRoomService.addEventListener('click', function(e) {
     roomService.removeAttribute('hidden');
     createRoomForm.setAttribute('hidden', 'createRoomForm');
-  })
+  });
+}
+
+if (mainBackToRoomService) {
+  mainBackToRoomService.addEventListener('click', function(e) {
+    roomService.removeAttribute('hidden');
+    container.setAttribute('hidden', 'container');
+  });
 }
 
 // Create room
@@ -119,7 +164,7 @@ if (createButton) {
   createButton.addEventListener('click', function(e) {
     createRoomForm.removeAttribute('hidden');
     roomService.setAttribute('hidden', 'roomService');
-  })
+  });
 }
 
 if (finalizeRoomButton) {
@@ -136,11 +181,11 @@ if (finalizeRoomButton) {
         db.child(name).set(pass);
         passList.child(name).set(pass);
         createRoomForm.setAttribute('hidden', 'createRoomForm');
-        chatArea.innerHTML = "Chat";
-       }
+        chatArea.innerHTML = 'Chat';
+      }
     });
-  })
-}
+  });
+};
 
 // Join room
 
@@ -148,26 +193,27 @@ if (joinRoom) {
   joinRoom.addEventListener('click', function(e) {
     joinForm.removeAttribute('hidden');
     roomService.setAttribute('hidden', 'roomService');
-  })
+  });
 }
 
+if (backToRoomServiceJoin) {
   backToRoomServiceJoin.addEventListener('click', function(e) {
     joinForm.setAttribute('hidden', 'joinForm');
     roomService.removeAttribute('hidden');
-  })
+  });
+}
 
 if (finalizeJoin) {
   finalizeJoin.addEventListener('click', function(e) {
-    joinName = roomNameJoin.value;
+    name = roomNameJoin.value;
     joinPass = roomPassJoin.value;
 
     roomToJoin = firebase.database().ref('passwords');
     roomToJoin.on("value", function(snapShot) {
 
 
-      if (snapShot.child(joinName).val() == joinPass) {
-        alert('after if');
-      databaseRoom.once('value')
+    if (snapShot.child(name).val() == joinPass) {
+      fieldRef.once('value')
         .then(function(snapshot) {
             snapshot.forEach(function(childSnap) {
               var p = document.createElement('p');
@@ -177,8 +223,8 @@ if (finalizeJoin) {
           });
         container.removeAttribute('hidden');
         joinForm.setAttribute('hidden', 'joinForm');
-	      alert('done');
+	      //mainText.value = '';
       }
     });
-  })
+  });
 }
